[Fix] fix an error

+++ src/main/java/com/example/breadbook/Report/ReportRepository.java
@@ -1,6 +1,7 @@
 package com.example.breadbook.Report;
 
+import com.example.breadbook.Report.model.Report;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReportRepository extends JpaRepository<Report, Integer> {
+public interface ReportRepository extends JpaRepository<Report, Long> {
 }
